Rename back
We don't need to rename it because we don't have a server reference in the client and client reference in the server.
